Improved coverage, actually runs trigger task now.
