+++ src/main/java/com/hmsonline/cassandra/triggers/CassandraStore.java
@@ -1,14 +1,19 @@
 package com.hmsonline.cassandra.triggers;
 
+import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.List;
 
 import org.apache.cassandra.thrift.Cassandra;
 import org.apache.cassandra.thrift.CassandraServer;
 import org.apache.cassandra.thrift.CfDef;
+import org.apache.cassandra.thrift.Column;
 import org.apache.cassandra.thrift.ColumnDef;
+import org.apache.cassandra.thrift.ColumnOrSuperColumn;
 import org.apache.cassandra.thrift.IndexType;
 import org.apache.cassandra.thrift.KsDef;
+import org.apache.cassandra.thrift.Mutation;
+import org.apache.cassandra.utils.ByteBufferUtil;
 import org.apache.commons.lang.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -20,12 +25,12 @@ public class CassandraStore {
     private String columnFamily = null;
 
     protected CassandraStore(String keyspace, String columnFamily) throws Exception {
-      this.keyspace = keyspace;
-      this.columnFamily = columnFamily;
-      this.create(new String [] {});
+        this.keyspace = keyspace;
+        this.columnFamily = columnFamily;
+        this.create(new String[] {});
     }
-    
-    protected CassandraStore(String keyspace, String columnFamily, String [] indexedColumns) throws Exception {
+
+    protected CassandraStore(String keyspace, String columnFamily, String[] indexedColumns) throws Exception {
         this.keyspace = keyspace;
         this.columnFamily = columnFamily;
         this.create(indexedColumns);
@@ -39,7 +44,7 @@ public class CassandraStore {
         return server;
     }
 
-    public synchronized void create(String [] indexedColumns) throws Exception {
+    public synchronized void create(String[] indexedColumns) throws Exception {
         if (!initialized) {
             try {
                 List<CfDef> cfDefList = new ArrayList<CfDef>();
@@ -55,26 +60,27 @@ public class CassandraStore {
                 columnFamily.setKey_validation_class("UTF8Type");
                 columnFamily.setDefault_validation_class("UTF8Type");
                 columnFamily.setComparator_type("UTF8Type");
-                
+
                 // add indexes on columns
                 if (indexedColumns != null && indexedColumns.length > 0) {
-                  for (Object indexedColumn : indexedColumns) {
-                      if (indexedColumn != null) {
-                          String indexedColumnStr = indexedColumn.toString();
-                          if (StringUtils.isNotBlank(indexedColumnStr)) {
-                              List<ColumnDef> columnMetadata = columnFamily.getColumn_metadata();
-                              columnMetadata = columnMetadata != null ? columnMetadata : new ArrayList<ColumnDef>();
-                              ColumnDef colDef = new ColumnDef();
-                              colDef.setName(indexedColumnStr.getBytes());
-                              colDef.index_type = IndexType.KEYS;
-                              colDef.setIndex_name(keyspace + "_" + this.columnFamily + "_" + indexedColumnStr + "_INDEX");
-                              columnMetadata.add(colDef);
-                              columnFamily.setColumn_metadata(columnMetadata);
-                          }
-                      }
-                  }
-              }
-                
+                    for (Object indexedColumn : indexedColumns) {
+                        if (indexedColumn != null) {
+                            String indexedColumnStr = indexedColumn.toString();
+                            if (StringUtils.isNotBlank(indexedColumnStr)) {
+                                List<ColumnDef> columnMetadata = columnFamily.getColumn_metadata();
+                                columnMetadata = columnMetadata != null ? columnMetadata : new ArrayList<ColumnDef>();
+                                ColumnDef colDef = new ColumnDef();
+                                colDef.setName(indexedColumnStr.getBytes());
+                                colDef.index_type = IndexType.KEYS;
+                                colDef.setIndex_name(keyspace + "_" + this.columnFamily + "_" + indexedColumnStr
+                                        + "_INDEX");
+                                columnMetadata.add(colDef);
+                                columnFamily.setColumn_metadata(columnMetadata);
+                            }
+                        }
+                    }
+                }
+
                 getConnection(this.getKeyspace()).system_add_column_family(columnFamily);
                 initialized = true;
             } catch (Exception e) {
@@ -88,16 +94,34 @@ public class CassandraStore {
         return keyspace;
     }
 
-    public void setKeyspace(String keyspace) {
-        this.keyspace = keyspace;
-    }
-
     public String getColumnFamily() {
         return columnFamily;
     }
-
-    public void setColumnFamily(String columnFamily) {
-        this.columnFamily = columnFamily;
+    
+    // Utility Methods
+    protected Mutation getMutation(String name, String value) {
+        return getMutation(name, ByteBufferUtil.bytes(value));
     }
 
+
+    protected Mutation getMutation(String name, ByteBuffer value) {
+        return getMutation(ByteBufferUtil.bytes(name), value);
+    }
+
+    protected Mutation getMutation(ByteBuffer name, OperationType value) {
+        return getMutation(name, ByteBufferUtil.bytes(value.toString()));
+    }
+
+    protected Mutation getMutation(ByteBuffer name, ByteBuffer value) {
+        Column c = new Column();
+        c.setName(name);
+        c.setValue(value);
+        c.setTimestamp(System.currentTimeMillis() * 1000);
+
+        Mutation m = new Mutation();
+        ColumnOrSuperColumn cc = new ColumnOrSuperColumn();
+        cc.setColumn(c);
+        m.setColumn_or_supercolumn(cc);
+        return m;
+    }
 }
